+++ tests/Unit/PigeonManagerTest.php
@@ -2,9 +2,12 @@
 
 namespace Convenia\Pigeon\Tests\Unit;
 
+use BadMethodCallException;
+use Convenia\Pigeon\Drivers\RabbitDriver;
 use Convenia\Pigeon\Exceptions\Driver\NullDriverException;
 use Convenia\Pigeon\PigeonManager;
 use Convenia\Pigeon\Tests\TestCase;
+use Mockery;
 
 class PigeonManagerTest extends TestCase
 {
@@ -76,5 +79,33 @@ class PigeonManagerTest extends TestCase
             'foo' => 'fighters',
         ]);
         $this->assertEquals($headers, $this->app['config']->get('pigeon.headers'));
+    }
+
+    public function test_it_should_delegate_calls_to_driver()
+    {
+        $driver = Mockery::mock(RabbitDriver::class);
+        $driver->shouldReceive('getChannel')
+            ->once()
+            ->with(28);
+
+        $manager = Mockery::mock(PigeonManager::class)->makePartial();
+        $manager->shouldReceive('driver')
+            ->twice()
+            ->andReturn($driver);
+
+        $manager->getChannel(28);
+
+        $this->assertFalse(
+            method_exists($manager, 'getChannel')
+        );
+    }
+
+    public function test_it_should_throw_exception_delegating_unimplemented_method()
+    {
+        $this->expectException(BadMethodCallException::class);
+        $this->expectExceptionMessage('Call to undefined method Convenia\Pigeon\PigeonManager::unimplementedMethod()');
+
+        $manager = new PigeonManager($this->app);
+        $manager->unimplementedMethod();
     }
 }